# Conflicts:
#	app/src/main/java/es/unex/giiis/golaso/adapters/ClasificacionAdapter.java
#	app/src/main/java/es/unex/giiis/golaso/adapters/JugadorEquipoAdapter.java
#	app/src/main/java/es/unex/giiis/golaso/adapters/ResultadoEquipoAdapter.java
#	app/src/main/java/es/unex/giiis/golaso/elementos/EquipoDetailFragment.java
#	app/src/main/java/es/unex/giiis/golaso/elementos/JugadoresFragment_equipo.java
#	app/src/main/java/es/unex/giiis/golaso/elementos/ResultadosFragment_equipo.java
#	app/src/main/java/es/unex/giiis/golaso/model/Equipo.java
#	app/src/main/java/es/unex/giiis/golaso/model/Jugador.java
#	app/src/main/java/es/unex/giiis/golaso/ui/clasificacion/ClasificacionFragment_clasificacion.java
#	app/src/main/res/layout/fragment_equipo_detail.xml
#	app/src/main/res/layout/fragment_jugadores_equipo.xml
#	app/src/main/res/layout/fragment_resultados_equipo.xml

+++ app/src/main/java/es/unex/giiis/golaso/adapters/ResultadoEquipoAdapter.java
@@ -28,9 +28,9 @@ public class ResultadoEquipoAdapter extends RecyclerView.Adapter<ResultadoEquipo
     List<Equipo> equipos;
     List<Partido> partidosEquipo;
     Equipo equipo;
-    int idEquipo;
+    long idEquipo;
 
-    public ResultadoEquipoAdapter(Context context, List<Partido> partidosList, List<Equipo> equiposList, int id) {
+    public ResultadoEquipoAdapter(Context context, List<Partido> partidosList, List<Equipo> equiposList, long id) {
         this.context = context;
         this.partidos = partidosList;
         this.equipos = equiposList;
@@ -134,7 +134,7 @@ public class ResultadoEquipoAdapter extends RecyclerView.Adapter<ResultadoEquipo
         return partidosEquipo.size();
     }
 
-    private Equipo buscarEquipoPorId(int id) {
+    private Equipo buscarEquipoPorId(long id) {
 
         boolean enc = false;
         Equipo e = null;
